Add spring-tool, a small utility to plot spring behavior
The weston_spring is a very flexible and powerful mecanhism for driving
animations.  However, it can be a little difficult to tame, but this
little helper can plot the response of the spring to a set of initial
parameters and makes it easy to tune and tweak the spring behavior.